move batch logic to general utils
