+++ quantem/core/utils/imaging_utils.py
@@ -1,9 +1,12 @@
 # Utilities for processing images
 
-from typing import Iterator, List, Optional, Tuple
+from typing import Optional, Tuple
 
 import numpy as np
+from numpy.typing import NDArray
 from scipy.ndimage import gaussian_filter
+
+from quantem.core.utils.utils import generate_batches
 
 
 def cross_correlation_shift(
@@ -146,26 +149,26 @@ def cross_correlation_shift(
 
 
 def bilinear_kde(
-    xa: np.ndarray,
-    ya: np.ndarray,
-    intensities: np.ndarray,
+    xa: NDArray,
+    ya: NDArray,
+    values: NDArray,
     output_shape: Tuple[int, int],
     kde_sigma: float,
     pad_value: float = 0.0,
     threshold: float = 1e-3,
     lowpass_filter: bool = False,
     max_batch_size: Optional[int] = None,
-) -> np.ndarray:
+) -> NDArray:
     """
     Compute a bilinear kernel density estimate (KDE) with smooth threshold masking.
 
     Parameters
     ----------
-    xa : np.ndarray
+    xa : NDArray
         Vertical (row) coordinates of input points.
-    ya : np.ndarray
+    ya : NDArray
         Horizontal (col) coordinates of input points.
-    intensities : np.ndarray
+    values : NDArray
         Weights for each (xa, ya) point.
     output_shape : tuple of int
         Output image shape (rows, cols).
@@ -182,7 +185,7 @@ def bilinear_kde(
 
     Returns
     -------
-    np.ndarray
+    NDArray
         The estimated KDE image with threshold-masked output.
     """
     rows, cols = output_shape
@@ -190,7 +193,7 @@ def bilinear_kde(
     yF = np.floor(ya.ravel()).astype(int)
     dx = xa.ravel() - xF
     dy = ya.ravel() - yF
-    w = intensities.ravel()
+    w = values.ravel()
 
     pix_count = np.zeros(rows * cols, dtype=np.float32)
     pix_output = np.zeros(rows * cols, dtype=np.float32)
@@ -236,74 +239,3 @@ def bilinear_kde(
         image = np.real(np.fft.ifft2(f_img))
 
     return image
-
-
-def subdivide_batches(
-    num_items: int,
-    num_batches: Optional[int] = None,
-    max_batch: Optional[int] = None,
-) -> List[int]:
-    """
-    Split `num_items` into a list of batch sizes.
-
-    Parameters
-    ----------
-    num_items : int
-        Total number of items to split into batches.
-    num_batches : int, optional
-        Number of desired batches. Cannot be used with `max_batch`.
-    max_batch : int, optional
-        Maximum batch size. Cannot be used with `num_batches`.
-
-    Returns
-    -------
-    List[int]
-        List of batch sizes that sum to `num_items`.
-    """
-    if num_batches is not None and max_batch is not None:
-        raise RuntimeError("Specify only one of `num_batches` or `max_batch`.")
-
-    if num_batches is None:
-        if max_batch is None:
-            raise RuntimeError("Must provide either `num_batches` or `max_batch`.")
-        num_batches = (num_items + max_batch - 1) // max_batch
-
-    if num_items < num_batches:
-        raise ValueError("`num_batches` may not exceed `num_items`.")
-
-    base_size = num_items // num_batches
-    remainder = num_items % num_batches
-
-    return [base_size + 1] * remainder + [base_size] * (num_batches - remainder)
-
-
-def generate_batches(
-    num_items: int,
-    num_batches: Optional[int] = None,
-    max_batch: Optional[int] = None,
-    start_index: int = 0,
-) -> Iterator[Tuple[int, int]]:
-    """
-    Yield (start, end) index tuples for each batch.
-
-    Parameters
-    ----------
-    num_items : int
-        Total number of items to batch.
-    num_batches : int, optional
-        Number of batches. Cannot be used with `max_batch`.
-    max_batch : int, optional
-        Maximum size of each batch. Cannot be used with `num_batches`.
-    start_index : int, default = 0
-        Optional offset to start indexing from.
-
-    Yields
-    ------
-    (int, int)
-        Tuple of (start, end) indices for each batch.
-    """
-    batch_sizes = subdivide_batches(num_items, num_batches, max_batch)
-    idx = start_index
-    for size in batch_sizes:
-        yield idx, idx + size
-        idx += size
