+++ quantem/core/utils/utils.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Iterator, List, Optional, Tuple
 
 import numpy as np
 from numpy.typing import NDArray
@@ -57,3 +57,74 @@ def as_numpy(array: Any) -> np.ndarray:
     if isinstance(array, np.ndarray):
         return array
     return np.asarray(array)
+
+
+def subdivide_batches(
+    num_items: int,
+    num_batches: Optional[int] = None,
+    max_batch: Optional[int] = None,
+) -> List[int]:
+    """
+    Split `num_items` into a list of batch sizes.
+
+    Parameters
+    ----------
+    num_items : int
+        Total number of items to split into batches.
+    num_batches : int, optional
+        Number of desired batches. Cannot be used with `max_batch`.
+    max_batch : int, optional
+        Maximum batch size. Cannot be used with `num_batches`.
+
+    Returns
+    -------
+    List[int]
+        List of batch sizes that sum to `num_items`.
+    """
+    if num_batches is not None and max_batch is not None:
+        raise RuntimeError("Specify only one of `num_batches` or `max_batch`.")
+
+    if num_batches is None:
+        if max_batch is None:
+            raise RuntimeError("Must provide either `num_batches` or `max_batch`.")
+        num_batches = (num_items + max_batch - 1) // max_batch
+
+    if num_items < num_batches:
+        raise ValueError("`num_batches` may not exceed `num_items`.")
+
+    base_size = num_items // num_batches
+    remainder = num_items % num_batches
+
+    return [base_size + 1] * remainder + [base_size] * (num_batches - remainder)
+
+
+def generate_batches(
+    num_items: int,
+    num_batches: Optional[int] = None,
+    max_batch: Optional[int] = None,
+    start_index: int = 0,
+) -> Iterator[Tuple[int, int]]:
+    """
+    Yield (start, end) index tuples for each batch.
+
+    Parameters
+    ----------
+    num_items : int
+        Total number of items to batch.
+    num_batches : int, optional
+        Number of batches. Cannot be used with `max_batch`.
+    max_batch : int, optional
+        Maximum size of each batch. Cannot be used with `num_batches`.
+    start_index : int, default = 0
+        Optional offset to start indexing from.
+
+    Yields
+    ------
+    (int, int)
+        Tuple of (start, end) indices for each batch.
+    """
+    batch_sizes = subdivide_batches(num_items, num_batches, max_batch)
+    idx = start_index
+    for size in batch_sizes:
+        yield idx, idx + size
+        idx += size
